feat(filter.c): internalizes services of the filters
